CI: Add powerset clippy check (#112)
* CI: Add powerset clippy check

As pointed out in #108, we don't explicitly check for all combinations
of features in the sdk, and more specifically, we miss warnings.

Add a CI step to run `cargo clippy -- --deny=warnings` on the powerset
of features in the sdk. Currently, this means 888 different crate
builds, but most of them are quite short.

Let's see how long it takes to run in CI to see if it's worth adding.

Also, fix some little issues that came up while running this.

* Oops, also install clippy

* Also fixup the name of the cache key

* Remove the bincode requirement for `SlotHashesSysvar`

* Fixup signature error